# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Forge 软件发布管理平台 API
  description: 本文档提供了 Forge 软件发布管理平台的详细 API 文档，包括认证、软件管理、统计分析等各个模块的接口说明。
  version: 1.0.0
  contact:
    name: Forge 开发团队
    email: dev@forge.com
servers:
  - url: http://localhost:5000/api
    description: 开发服务器
  - url: https://api.forge.com/api
    description: 生产服务器
security:
  - BearerAuth: []
paths:
...
  /statistics/overview:
    get:
      tags:
        - 统计分析
      summary: 获取统计概览
      description: 获取系统的整体统计信息
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取统计概览成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 获取统计概览成功
                  data:
                    type: object
                    properties:
                      spaces_count:
                        type: integer
                        description: 软件空间总数
                        example: 10
                      versions_count:
                        type: integer
                        description: 软件版本总数
                        example: 25
                      published_versions_count:
                        type: integer
                        description: 已发布版本数
                        example: 20
                      total_downloads:
                        type: integer
                        description: 总下载次数
                        example: 1000
                      recent_downloads:
                        type: integer
                        description: 最近7天下载次数
                        example: 100
                      monthly_downloads:
                        type: integer
                        description: 最近30天下载次数
                        example: 300
                      unique_ips:
                        type: integer
                        description: 独立IP下载用户数
                        example: 500
                      spaces_downloads:
                        type: array
                        description: 按软件空间统计的下载次数
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                              description: 软件空间名称
                              example: Forge
                            downloads:
                              type: integer
                              description: 下载次数
                              example: 100
                      versions_downloads:
                        type: array
                        description: 按版本统计的下载次数
                        items:
                          type: object
                          properties:
                            space_name:
                              type: string
                              description: 软件空间名称
                              example: Forge
                            version:
                              type: string
                              description: 版本号
                              example: 1.0.0
                            downloads:
                              type: integer
                              description: 下载次数
                              example: 50
                      recent_downloads_info:
                        type: array
                        description: 最近下载记录
                        items:
                          $ref: '#/components/schemas/DownloadRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /statistics/downloads:
    get:
      tags:
        - 统计分析
      summary: 获取下载统计
      description: 获取下载记录的统计信息
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 20
        - name: space_id
          in: query
          description: 软件空间ID
          schema:
            type: integer
        - name: version_id
          in: query
          description: 版本ID
          schema:
            type: integer
        - name: start_date
          in: query
          description: 开始日期（YYYY-MM-DD）
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: 结束日期（YYYY-MM-DD）
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 获取下载统计成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 获取下载统计成功
                  data:
                    type: object
                    properties:
                      downloads:
                        type: array
                        description: 下载记录列表
                        items:
                          $ref: '#/components/schemas/DownloadRecord'
                      total:
                        type: integer
                        description: 总记录数
                        example: 100
                      pages:
                        type: integer
                        description: 总页数
                        example: 5
                      current_page:
                        type: integer
                        description: 当前页码
                        example: 1
                      per_page:
                        type: integer
                        description: 每页数量
                        example: 20
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /statistics/downloads/timeline:
    get:
      tags:
        - 统计分析
      summary: 获取下载时间线统计
      description: 获取指定时间范围内的下载时间线统计
      security:
        - BearerAuth: []
      parameters:
        - name: days
          in: query
          description: 统计天数（最多365天）
          schema:
            type: integer
            default: 30
            maximum: 365
        - name: space_id
          in: query
          description: 软件空间ID
          schema:
            type: integer
      responses:
        '200':
          description: 获取下载时间线统计成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 获取下载时间线统计成功
                  data:
                    type: object
                    properties:
                      timeline:
                        type: array
                        description: 时间线数据
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                              format: date
                              description: 日期
                              example: 2023-01-01
                            downloads:
                              type: integer
                              description: 下载次数
                              example: 50
                      start_date:
                        type: string
                        format: date
                        description: 开始日期
                        example: 2023-01-01
                      end_date:
                        type: string
                        format: date
                        description: 结束日期
                        example: 2023-01-30
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /statistics/spaces/{space_id}:
    get:
      tags:
        - 统计分析
      summary: 获取软件空间统计信息
      description: 获取指定软件空间的统计信息
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SpaceId'
      responses:
        '200':
          description: 获取软件空间统计信息成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 获取软件空间统计信息成功
                  data:
                    type: object
                    properties:
                      space_info:
                        type: object
                        description: 软件空间基本信息
                        properties:
                          id:
                            type: integer
                            description: 软件空间ID
                            example: 1
                          name:
                            type: string
                            description: 软件名称
                            example: Forge
                          description:
                            type: string
                            description: 软件描述
                            example: 一个强大的软件发布管理平台
                          author:
                            type: string
                            description: 作者
                            example: Forge Team
                          created_at:
                            type: string
                            format: date-time
                            description: 创建时间
                            example: 2023-01-01T00:00:00Z
                          versions_count:
                            type: integer
                            description: 版本数量
                            example: 5
                          downloads_count:
                            type: integer
                            description: 下载次数
                            example: 1000
                      versions_downloads:
                        type: array
                        description: 各版本下载次数
                        items:
                          type: object
                          properties:
                            version:
                              type: string
                              description: 版本号
                              example: 1.0.0
                            is_published:
                              type: boolean
                              description: 是否已发布
                              example: true
                            downloads:
                              type: integer
                              description: 下载次数
                              example: 500
                      recent_downloads:
                        type: array
                        description: 最近下载记录
                        items:
                          $ref: '#/components/schemas/DownloadRecord'
                      timeline:
                        type: array
                        description: 最近30天下载趋势
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                              format: date
                              description: 日期
                              example: 2023-01-01
                            downloads:
                              type: integer
                              description: 下载次数
                              example: 50
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Webhook相关接口
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token认证
  schemas:
...
    DownloadRecord:
      type: object
      properties:
        id:
          type: integer
          description: 下载记录ID
          example: 1
        space_name:
          type: string
          description: 软件空间名称
          example: Forge
        version:
          type: string
          description: 版本号
          example: 1.0.0
        ip_address:
          type: string
          description: IP地址
          example: 192.168.1.1
        user_agent:
          type: string
          description: 用户代理
          example: Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36
        download_time:
          type: string
          format: date-time
          description: 下载时间
          example: 2023-01-01T10:00:00Z
...
  parameters:
    SpaceId:
      name: space_id
      in: path
      required: true
      description: 软件空间ID
      schema:
        type: integer
...
  responses:
...
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 请求参数错误
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: INVALID_PARAMETER
                  message:
                    type: string
                    example: 请求参数无效
    Unauthorized:
      description: 未授权
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 未授权，请登录
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: UNAUTHORIZED
                  message:
                    type: string
                    example: 未授权访问
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 禁止访问
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: FORBIDDEN
                  message:
                    type: string
                    example: 权限不足
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 资源不存在
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: RESOURCE_NOT_FOUND
                  message:
                    type: string
                    example: 资源不存在
...